requirements.txt
# AIFX - AI-Enhanced Forex Trading System Dependencies
# AIFX - AI增強外匯交易系統依賴項
# Compatible versions for Python 3.12 | 適用於Python 3.12的兼容版本

# Core Data Processing | 核心數據處理
pandas>=1.5.0,<3.0.0
numpy>=1.21.0,<2.0.0
scipy>=1.7.0

# Financial Data & Technical Analysis | 金融數據與技術分析
yfinance>=0.2.0              # Yahoo Finance data download | Yahoo財經數據下載
mplfinance                   # Financial plotting (latest stable) | 金融圖表（最新穩定版）

# Machine Learning | 機器學習  
scikit-learn>=1.0.0          # ML algorithms and utilities | ML演算法和工具
xgboost>=1.6.0               # Gradient boosting framework | 梯度提升框架
lightgbm>=3.3.0              # Light gradient boosting | 輕量級梯度提升

# Deep Learning (Phase 2 Addition) | 深度學習（第二階段新增）
tensorflow>=2.10.0           # Deep learning framework for LSTM | 用於LSTM的深度學習框架
keras>=2.10.0                # High-level neural networks API | 高級神經網絡API

# Data Visualization | 數據可視化
matplotlib>=3.5.0            # Basic plotting | 基礎繪圖
seaborn>=0.11.0              # Statistical visualization | 統計可視化
plotly>=5.10.0               # Interactive plotting | 互動式繪圖

# Development & Testing | 開發與測試
pytest>=7.0.0               # Testing framework | 測試框架
pytest-cov>=4.0.0           # Coverage testing | 覆蓋率測試

# Configuration & Utilities | 配置與工具
pyyaml>=6.0                  # YAML configuration | YAML配置
requests>=2.25.0             # HTTP requests | HTTP請求
tqdm>=4.60.0                 # Progress bars | 進度條
joblib>=1.1.0                # Model serialization | 模型序列化
jsonschema>=4.0.0            # JSON schema validation | JSON模式驗證

# Database Support | 資料庫支援
sqlalchemy>=2.0.0            # SQL toolkit and ORM | SQL工具包和ORM
pyodbc>=4.0.39               # SQL Server driver | SQL Server驅動程式
redis>=4.5.0                 # Redis client for caching | Redis緩存客戶端
psycopg2-binary>=2.9.0       # PostgreSQL adapter | PostgreSQL適配器

# Time & Date Handling | 時間與日期處理  
pytz>=2021.1                 # Timezone handling | 時區處理
python-dateutil>=2.8.0       # Date utilities | 日期工具

# Logging & Monitoring | 日誌與監控
structlog>=22.0.0            # Structured logging | 結構化日誌

# Real-time Data Pipeline (Phase 4.2) | 實時數據管道（第四階段第二節）
websocket-client>=1.5.0      # WebSocket client for real-time data | 實時數據WebSocket客戶端
aiohttp>=3.8.0               # Async HTTP client | 異步HTTP客戶端
python-socketio>=5.8.0       # Socket.IO support | Socket.IO支援
asyncio>=3.4.3               # Async programming support | 異步程式支援

# Web Interface (merged from requirements-web.txt) | 網頁界面（從requirements-web.txt合併）
fastapi>=0.104.0             # Modern web framework | 現代網頁框架
uvicorn[standard]>=0.24.0    # ASGI server | ASGI服務器
websockets>=12.0             # WebSocket support | WebSocket支援
jinja2>=3.1.0                # Template engine | 模板引擎
python-multipart>=0.0.6      # Multipart form support | 多部分表單支援
aiofiles>=23.2.0             # Async file operations | 異步文件操作
psutil>=5.9.0                # System monitoring | 系統監控

# Optional Advanced Packages (install separately if needed) | 可選高級包（需要時單獨安裝）
# tensorflow>=2.10.0         # Deep learning framework | 深度學習框架
# torch>=1.12.0              # PyTorch deep learning | PyTorch深度學習  
# backtrader>=1.9.78.123     # Backtesting framework | 回測框架
# jupyter>=1.0.0             # Jupyter notebook | Jupyter筆記本
# ta-lib>=0.4.0              # Technical Analysis Library | 技術分析庫
# pandas-ta>=0.3.14b0        # Technical analysis indicators | 技術分析指標